+++ src/config.default.ts
@@ -953,15 +953,3 @@ export function parseDiscloseVersion(raw: string | undefined): boolean {
 
 export const SPEC_URL =
   'https://digitalcredentials.github.io/wallet-attached-storage-spec/'
-export const UBC_MANIFEST_URL =
-  'https://codeberg.org/fediverse/fep/src/branch/main/fep/6fcd/fep-6fcd.md#manifest-file'
-export const SPACE_URL =
-  'https://digitalcredentials.github.io/wallet-attached-storage-spec/#spaces'
-export const COLLECTION_URL =
-  'https://digitalcredentials.github.io/wallet-attached-storage-spec/#collection-data-model'
-export const RESOURCE_URL =
-  'https://digitalcredentials.github.io/wallet-attached-storage-spec/#resource-data-model'
-export const POLICY_URL =
-  'https://digitalcredentials.github.io/wallet-attached-storage-spec/#policy'
-export const META_URL =
-  'https://digitalcredentials.github.io/wallet-attached-storage-spec/#resource-metadata-data-model'